refactor(rpc-alt): factor out sui-name-service
## Description
Pull out `sui-name-service` from `sui-json-rpc`, because it is used in
multiple other crates. This change also re-orders the contents of the
`name_service` module and brings its tests into one file.

## Test plan
CI
