+++ crates/sui-indexer/Cargo.toml
@@ -54,6 +54,7 @@ sui-json.workspace = true
 sui-json-rpc.workspace = true
 sui-json-rpc-api.workspace = true
 sui-json-rpc-types.workspace = true
+sui-name-service.workspace = true
 sui-open-rpc.workspace = true
 sui-pg-db.workspace = true
 sui-sdk.workspace = true
