+++ crates/sui-json-rpc/src/lib.rs
@@ -48,7 +48,6 @@ pub mod indexer_api;
 pub mod logger;
 mod metrics;
 pub mod move_utils;
-pub mod name_service;
 mod object_changes;
 pub mod read_api;
 mod traffic_control;
