+++ crates/sui-mvr-graphql-rpc/tests/move_registry_e2e.rs
@@ -5,7 +5,6 @@ use std::str::FromStr;
 use std::{path::PathBuf, time::Duration};
 
 use sui_graphql_rpc_client::simple_client::SimpleClient;
-use sui_json_rpc::name_service::{Domain, DomainFormat};
 use sui_json_rpc_types::ObjectChange;
 use sui_move_build::BuildConfig;
 use sui_mvr_graphql_rpc::{
@@ -15,6 +14,7 @@ use sui_mvr_graphql_rpc::{
         wait_for_graphql_checkpoint_catchup, wait_for_graphql_server, NetworkCluster,
     },
 };
+use sui_name_service::{Domain, DomainFormat};
 use sui_pg_db::temp::get_available_port;
 use sui_types::{
     base_types::{ObjectID, SequenceNumber},
